chore: code cleanup and add back batch tests
